cleanup, API for starting a game
